net: netsocks
The user may now call socket() in order to setup a TCP socket, which
reads and writes to the TCP/IP stack on a seperate network thread.

The client asks for a new socket() using the netctrl interface, which
sends a NC_SOCKET command into the netctrl queue, which then the network
thread will read and respond to creating a new TCP control block.

Then, the client may use NC_CONNECT to connect to a specific host:port
pair. Further, basic net_read & net_write functions can be called to
communicate with the remote host.

The net_read & net_write functions push and pull data to and from the
rbuf & wbuf queues in the netsock struct - the network thread
periodically checks both queues, moving data from the TCP/IP stack into
the netsock bufs and reversed.