net: bind & accept
Binding a socket to listen for connections is now possible with the
NC_BIND netctrl command. After success, the user can issue an accept()
call which blocks until the network thread finds a client that wants to
connect.

This is done with another queue, namely the waiting_client single-slot
queue. Accept() waits for the queue to be filled with a new netsock by
the netsock_tcp_accept callback.