net: stop wireless interface from disconnecting itself
This is a strange one: after some inactivity the wireless interface
could simply stop replying to any network traffic (say, ARP requests).
Because of this, other hosts could no longer ping or open TCP ports
to the device.

If we poll the cyw43 driver for RSSI each time we loop, it seems to
keep the connection open. Weird.