meta: create projects differently
This new system allows for creating independant projects, only using
parts of the whole micron system. A project is defined as a directory
in the dist/projects directory with a project.json declaring the used
source files, libraries, board & clangd settings.

Before building, selecting a project is required (this can be done with
the new dist/configure script).